refactor: Features/Exam SPM 패키지 연결, QRIZ/Feature/Exam/ 삭제
