--- QRIZ/Feature/Exam/ExamResult/HostingController/ExamResultHostingController.swift
@@ -1,12 +1,0 @@
-//
-//  ExamResultHostingController.swift
-//  QRIZ
-//
-//  Created by 이창현 on 5/26/25.
-//
-
-import UIKit
-import SwiftUI
-
-final class ExamResultHostingController: UIHostingController<ExamResultView> {
-}
